Move some code around to clean up a bit
Changes:
 - Thread::Temporary is now named PerThreadTemporary (in accumulators.hpp)
 - Context::data is now a separate class PerContextAccumulators (in accumulators.hpp)
 - Metric::finalize is now a method on PerThreadTemporary

No functional changes.
