+++ src/lib/profile/Makefile.in
@@ -157,6 +157,7 @@ am__objects_1 = stdshim/libHPCprofile_la-shared_mutex.lo \
 	libHPCprofile_la-metric.lo libHPCprofile_la-module.lo \
 	libHPCprofile_la-attributes.lo libHPCprofile_la-dataclass.lo \
 	libHPCprofile_la-sink.lo libHPCprofile_la-packedids.lo \
+	libHPCprofile_la-accumulators.lo \
 	sources/libHPCprofile_la-packed.lo \
 	sources/libHPCprofile_la-hpcrun4.lo \
 	sinks/libHPCprofile_la-hpctracedb2.lo \
@@ -577,6 +578,7 @@ MYSOURCES = \
 	\
 	pipeline.cpp source.cpp context.cpp scope.cpp metric.cpp \
 	module.cpp attributes.cpp dataclass.cpp sink.cpp packedids.cpp \
+	accumulators.cpp \
 	\
 	sources/packed.cpp sources/hpcrun4.cpp \
 	\
@@ -772,6 +774,7 @@ mostlyclean-compile:
 distclean-compile:
 	-rm -f *.tab.c
 
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libHPCprofile_la-accumulators.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libHPCprofile_la-attributes.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libHPCprofile_la-context.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libHPCprofile_la-dataclass.Plo@am__quote@
@@ -991,6 +994,13 @@ libHPCprofile_la-packedids.lo: packedids.cpp
 @AMDEP_TRUE@@am__fastdepCXX_FALSE@	$(AM_V_CXX)source='packedids.cpp' object='libHPCprofile_la-packedids.lo' libtool=yes @AMDEPBACKSLASH@
 @AMDEP_TRUE@@am__fastdepCXX_FALSE@	DEPDIR=$(DEPDIR) $(CXXDEPMODE) $(depcomp) @AMDEPBACKSLASH@
 @am__fastdepCXX_FALSE@	$(AM_V_CXX@am__nodep@)$(LIBTOOL) $(AM_V_lt) --tag=CXX $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CXX) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libHPCprofile_la_CXXFLAGS) $(CXXFLAGS) -c -o libHPCprofile_la-packedids.lo `test -f 'packedids.cpp' || echo '$(srcdir)/'`packedids.cpp
+
+libHPCprofile_la-accumulators.lo: accumulators.cpp
+@am__fastdepCXX_TRUE@	$(AM_V_CXX)$(LIBTOOL) $(AM_V_lt) --tag=CXX $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CXX) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libHPCprofile_la_CXXFLAGS) $(CXXFLAGS) -MT libHPCprofile_la-accumulators.lo -MD -MP -MF $(DEPDIR)/libHPCprofile_la-accumulators.Tpo -c -o libHPCprofile_la-accumulators.lo `test -f 'accumulators.cpp' || echo '$(srcdir)/'`accumulators.cpp
+@am__fastdepCXX_TRUE@	$(AM_V_at)$(am__mv) $(DEPDIR)/libHPCprofile_la-accumulators.Tpo $(DEPDIR)/libHPCprofile_la-accumulators.Plo
+@AMDEP_TRUE@@am__fastdepCXX_FALSE@	$(AM_V_CXX)source='accumulators.cpp' object='libHPCprofile_la-accumulators.lo' libtool=yes @AMDEPBACKSLASH@
+@AMDEP_TRUE@@am__fastdepCXX_FALSE@	DEPDIR=$(DEPDIR) $(CXXDEPMODE) $(depcomp) @AMDEPBACKSLASH@
+@am__fastdepCXX_FALSE@	$(AM_V_CXX@am__nodep@)$(LIBTOOL) $(AM_V_lt) --tag=CXX $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CXX) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libHPCprofile_la_CXXFLAGS) $(CXXFLAGS) -c -o libHPCprofile_la-accumulators.lo `test -f 'accumulators.cpp' || echo '$(srcdir)/'`accumulators.cpp
 
 sources/libHPCprofile_la-packed.lo: sources/packed.cpp
 @am__fastdepCXX_TRUE@	$(AM_V_CXX)$(LIBTOOL) $(AM_V_lt) --tag=CXX $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CXX) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libHPCprofile_la_CXXFLAGS) $(CXXFLAGS) -MT sources/libHPCprofile_la-packed.lo -MD -MP -MF sources/$(DEPDIR)/libHPCprofile_la-packed.Tpo -c -o sources/libHPCprofile_la-packed.lo `test -f 'sources/packed.cpp' || echo '$(srcdir)/'`sources/packed.cpp
