+++ src/lib/profile/attributes.hpp
@@ -170,76 +170,6 @@ public:
 
   /// Attributes accociated with this Thread.
   ThreadAttributes attributes;
-
-  /// Thread-local data that is temporary and should be removed quickly.
-  class Temporary {
-  public:
-    // Access to the backing thread.
-    operator Thread&() noexcept { return m_thread; }
-    operator const Thread&() const noexcept { return m_thread; }
-    Thread& thread() noexcept { return m_thread; }
-    const Thread& thread() const noexcept { return m_thread; }
-
-    // Movable, not copiable
-    Temporary(const Temporary&) = delete;
-    Temporary(Temporary&&) = default;
-
-    /// Reference to the Metric data for a particular Context in this Thread.
-    /// Returns `std::nullopt` if none is present.
-    // MT: Safe (const), Unstable (before notifyThreadFinal)
-    auto accumulatorsFor(const Context& c) const noexcept {
-      return c_data.find(c);
-    }
-
-    /// Reference to all of the Metric data on Thread.
-    // MT: Safe (const), Unstable (before notifyThreadFinal)
-    const auto& accumulators() const noexcept { return c_data; }
-
-  private:
-    Thread& m_thread;
-
-    friend class ProfilePipeline;
-    Temporary(Thread& t) : m_thread(t) {};
-
-    // Bits needed for handling timepoints
-    std::chrono::nanoseconds minTime = std::chrono::nanoseconds::max();
-    std::chrono::nanoseconds maxTime = std::chrono::nanoseconds::min();
-    template<class Tp>
-    struct TimepointsData {
-      bool unboundedDisorder = false;
-      util::bounded_streaming_sort_buffer<Tp, util::compare_only_first<Tp>> sortBuf;
-      std::vector<Tp> staging;
-    };
-    TimepointsData<std::pair<std::chrono::nanoseconds,
-      std::reference_wrapper<const Context>>> ctxTpData;
-    util::locked_unordered_map<util::reference_index<const Metric>,
-      TimepointsData<std::pair<std::chrono::nanoseconds, double>>> metricTpData;
-
-    friend class Metric;
-    util::locked_unordered_map<util::reference_index<const Context>,
-      util::locked_unordered_map<util::reference_index<const Metric>,
-        MetricAccumulator>> c_data;
-    util::locked_unordered_map<util::reference_index<const ContextReconstruction>,
-      util::locked_unordered_map<util::reference_index<const Metric>,
-        MetricAccumulator>> r_data;
-
-    struct RGroup {
-      util::locked_unordered_map<util::reference_index<const Context>,
-        util::locked_unordered_map<util::reference_index<const Metric>,
-          MetricAccumulator>> c_data;
-      util::locked_unordered_map<util::reference_index<const ContextFlowGraph>,
-        util::locked_unordered_map<util::reference_index<const Metric>,
-          MetricAccumulator>> fg_data;
-
-      std::mutex lock;
-      std::unordered_map<Scope,
-        std::unordered_set<util::reference_index<Context>>> c_entries;
-      std::unordered_map<util::reference_index<ContextFlowGraph>,
-        std::unordered_set<util::reference_index<const ContextReconstruction>>>
-          fg_reconsts;
-    };
-    util::locked_unordered_map<uint64_t, RGroup> r_groups;
-  };
 };
 
 /// Attributes unique to a particular profile. Since a profile represents an
