+++ src/lib/profile/sinks/sparsedb.hpp
@@ -90,7 +90,7 @@ public:
   void notifyPipeline() noexcept override;
 
   void notifyWavefront(hpctoolkit::DataClass) noexcept override;
-  void notifyThreadFinal(const hpctoolkit::Thread::Temporary&) override;
+  void notifyThreadFinal(const hpctoolkit::PerThreadTemporary&) override;
 
   void cctdbSetUp();
   void writeCCTDB();
